Fixed the worker node's queries, as well as the .pro file for the sql tests.
Not entirely sure why the CI test cases aren't working, but they are working
locally (famous last words, right?). It appears to be an issue stemming from
a  lack of QMYSQL driver, which would make sense. It was even a pain to get
that driver up and running _locally_ (let alone remotely).